Refactor checkpoints controller in backend
