+++ backend/src/index.ts
@@ -2,7 +2,7 @@ import express from "express"
 import { PrismaClient } from "../prisma/prisma/"
 import cors from "cors"
 import path from "path"
-import checkpointsRouter from "../controllers/checkpoints"
+import checkpointsRouter from "./routes/checkpoints.route"
 import groupsRouter from "../controllers/groups"
 import { unknownEndpoint, errorHandler } from "../utils/middleware"
 import routesRouter from "./routes/routes.route"
